Group host dropout no longer an issue
ZMQ uses TCP batch messaging in its implementation. This means that while it
appears to be using one socket per connection through the API, that is not the
case. It batches messages through a fewer number of ports underneath the hood.

The problem arose out of calling close() on what was thought to be one of these
singular sockets. This, in fact, closed the underlying TCP socket and not the
ZMQ socket which caused all batched hosts to drop out as well when attempting
to reconnect. Using disconnect(), the literal opposite of connect(), gives the
desired result. A poller was unnecessary, contrary to previous thoughts.

Fixes #9.