Removed ZMQ as a Dependency
ZMQ was causing all hosts to drop in and out if one was trying to reconnect.
Thankfully, rbthomp demonstrated that this software could be used without ZMQ
if the correct miner config was used, which is now documented in the README.

miner-apid.py is no longer needed, but will remain in the repository for now,
unaltered (meaning it still has the ZMQ dependency).

Fixes #9.