Pool Mining and Solo Mining
Implemented a feature to allow the monitor to present information as if it were
solo mining or pool mining. This doesn't actually set miners to mine solo or as
a part of a pool. Available for configuration as a boolean in the script or as
a commandline option.

Also added help information as a command line option.

Updated the apply_formatting function to be a little neater.